% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_anaconda_tos.R
\name{accept_anaconda_tos}
\alias{accept_anaconda_tos}
\title{accept_anaconda_tos}
\usage{
accept_anaconda_tos(conda_bin = NULL)
}
\description{
Reorder factor levels within each topic panel.
Based on tidytext's reorder_within().
}
\keyword{internal}
