% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_classifications_table.R
\name{extract_classifications_table}
\alias{extract_classifications_table}
\title{extract_classifications_table}
\usage{
extract_classifications_table(file_path, max_records = NULL)
}
\arguments{
\item{file_path}{A path to a jsonl file exported from lens.org}

\item{max_records}{The maximum number of records to be extracted}
}
\value{
Returns a dataframe with: "lens_id"  , "Classification_system", "Symbol", "Classification_value", "Sequence"
classification_system: IPCR for the international patent classification system, or CPCR, for the cooperative patent classification system.
The field 'Symbol' gives the class.
The field 'classification_value' (for CPC or IPRC classifications) can take values like I or L. From the Lens docs:
I stands for “Invention”.
L stands for “Later”.
So if you see a classification entry with "classification_value": "I", it means that that classification symbol is flagged as relating to the invention as determined by the patent office/ Lens — i.e., the classification is considered part of the core inventive subject matter.
If it were "L", it would be a classification assigned later (for example: a supplementary, non-primary classification added after initial assignment) rather than necessarily part of the primary inventive content.
The field 'Sequence' gives the position of the symbol in the chain of classifications
}
\description{
This extracts information from a lens.org patent export
}
\examples{
example_file <- system.file("extdata", "lens-export-3.jsonl", package = "NetworkIsLifeR")
plop <- extract_classifications_table(example_file)
}
