% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_giant_component.R
\name{extract_giant_component}
\alias{extract_giant_component}
\title{Extract a Giant Component}
\usage{
extract_giant_component(graph)
}
\arguments{
\item{graph}{an igraph object}
}
\value{
The giant component of the provided graph
}
\description{
This function Extracts the giant component of a network
}
\examples{
# Create a random graph with multiple components
set.seed(123)
g <- igraph::sample_gnp(20, 0.1) # 20 nodes, edge probability 0.1
# Extract the giant component using your function
g_giant <- extract_giant_component(g)
}
