% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_key.R
\name{fingerprint_key_vec}
\alias{fingerprint_key_vec}
\title{Compute fingerprint keys for a character vector}
\usage{
fingerprint_key_vec(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of fingerprint keys the same length as \code{x},
with \code{NA} values preserved.
}
\description{
A vectorised wrapper around \code{\link[=fingerprint_key]{fingerprint_key()}} that returns a character
vector of fingerprint keys with the same length as the input.
}
\examples{
x <- c("Philips", "philips", "PHILIPS", "Sony", "SONY", NA)
fingerprint_key_vec(x)
# [1] "philips" "philips" "philips" "sony"    "sony"    NA

}
\seealso{
\code{\link[=fingerprint_key]{fingerprint_key()}} for the scalar version,
\code{\link[=fingerprint_clean]{fingerprint_clean()}} to apply clustering to a data frame column.
}
