% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topic_terms_grid.R
\name{plot_topic_terms_grid}
\alias{plot_topic_terms_grid}
\title{Plot top terms per topic as small barplots}
\usage{
plot_topic_terms_grid(
  topic_terms,
  n_topics = 6,
  n_terms = 10,
  metric = c("tfidf", "n"),
  topics = NULL,
  topic_labels = NULL
)
}
\arguments{
\item{topic_terms}{A data.frame/tibble with columns:
\itemize{
\item topic
\item term
\item tfidf (or another metric you choose via \code{metric})
}}

\item{n_topics}{How many topics to visualise (default: 6).
If there are fewer topics than this, all topics are shown.}

\item{n_terms}{How many top terms per topic (default: 10).}

\item{metric}{Which column to use for bar height (default: "tfidf").
You can set this to "n" if you want raw counts instead.}

\item{topics}{Optional vector of topic IDs to plot explicitly.
If provided, this overrides \code{n_topics}.}

\item{topic_labels}{A dataframe with two columns ("cluster" and "label") that
will put a label on the topics.}
}
\value{
A ggplot object with facetted barplots.
}
\description{
Plot top terms per topic as small barplots
}
