% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_within.R
\name{reorder_within}
\alias{reorder_within}
\title{reorder_within}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___")
}
\description{
Reorder factor levels within each topic panel.
Based on tidytext's reorder_within().
}
\keyword{internal}
