% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_bertopic_env.R
\name{setup_bertopic_env}
\alias{setup_bertopic_env}
\title{setup_bertopic_env}
\usage{
setup_bertopic_env(
  envname = "bertopic_r_env",
  python_version = "3.11",
  use_conda_on_windows = TRUE
)
}
\arguments{
\item{envname}{Name of the environment.}

\item{python_version}{Python version to install/use (default "3.11").}

\item{use_conda_on_windows}{Logical; if TRUE on Windows use conda+Miniconda,
otherwise fall back to virtualenv.}
}
\description{
This function checks for a working Python installation and installs one
if none is available. It then creates a dedicated Python virtual
environment and installs the Python packages required for using
\code{bertopicR}. If the environment already exists, the function exits
safely without making changes. This ensures that all Python
dependencies needed by \code{bertopicR} are available and isolated in a
controlled environment.
}
