% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_applicants_table.R
\name{extract_applicants_table}
\alias{extract_applicants_table}
\title{extract_applicants_table}
\usage{
extract_applicants_table(file_path, max_records = NULL)
}
\arguments{
\item{file_path}{A path to a jsonl file exported from lens.org}

\item{max_records}{The maximum number of records to be extracted}
}
\value{
Returns a dataframe with: "lens_id" ,  "sequence" , "name" ,     "residence", "address"
}
\description{
This extracts information about the applicants. It will export one line per applicant
}
\examples{
example_file <- system.file("extdata", "lens-export-3.jsonl", package = "NetworkIsLifeR")
extract_applicants_table(example_file)
}
