% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_patent_jsonl.R
\name{process_patent_jsonl}
\alias{process_patent_jsonl}
\title{process_patent_jsonl}
\usage{
process_patent_jsonl(file_path, max_records = NULL, verbose = TRUE)
}
\arguments{
\item{file_path}{A path to a jsonl file exported from lens.org.}

\item{max_records}{The maximum number of records to be extracted.}

\item{verbose}{Extra information return from the function.}
}
\value{
Returns a dataframe with: "lens_id"  ,            "jurisdiction"     ,    "doc_number" ,          "kind"
"date_published"  ,     "lang"  ,               "title"      ,          "abstract",
"app_doc_number"  ,     "app_date"   ,          "priority_date" ,        "applicants" ,
"inventors"    ,        "ipc_classifications",   "patent_status"    ,    "granted" ,
"grant_date"    ,       "cited_by_count" ,      "simple_family_size" ,  "extended_family_size"
}
\description{
This extracts information from a lens.org patent export
}
\examples{
library(purrr)
library(dplyr)
example_file <- system.file("extdata", "lens-export-3.jsonl", package = "NetworkIsLifeR")
process_patent_jsonl(example_file)
}
